requirements.txt
fastapi
uvicorn
python-dotenv
openai
gtts
pygame
speechrecognition
pydantic
requests